requirements.txt
numpy==1.26.3
pandas==2.1.4
python-dateutil==2.8.2
python-dotenv==1.0.0
spotipy==2.22.1
yarl==1.9.4
